+++ sdk/face/azure-ai-vision-face/CHANGELOG.md
@@ -1,14 +1,23 @@
 # Release History
 
-## 1.0.0-beta.3 (Unreleased)
+## 1.0.0-beta.3 (2025-07-11)
 
-### Features Added
 
 ### Breaking Changes
 
-### Bugs Fixed
+- Change the default service API version to `v1.2`
+- Removed `CreateLivenessSessionResult` and use `LivenessSession` for liveness session creation and get session.
+- Unified `CreateLivenessWithVerifySession` and require the verify image when creating a liveness with verify session.
+- Update model `LivenessSession`, `CreateLivenessSessionContent`, and `CreateLivenessWithVerifySessionContent`.
+- Removed `LivenessSessionAudit`
+- Removed `GetLivenessSessions` and `GetLivenessWithVerifySessions`
+- Removed `PersonDirectory` operations
 
 ### Other Changes
+
+- Removed FaceClient customizations
+- Liveness
+  - Added `LivenessSessionAttempt`
 
 ## 1.0.0-beta.2 (2024-10-23)